requirements.txt
flask
playwright
chromadb
python-dotenv
textstat
openai
